Split up test_tools.py
Tidy up Fourier transform and Hologram tests

Remove parallel from hologram tests since this is just passed straight to .map() anyway, which is tested elsewhere, and there is a risk of poor performance when parallel testing on CI VMs such as Travis/Azure

Bugfix
