+++ hyperspy/tests/signal/test_hologram_image.py
@@ -21,29 +21,57 @@ import gc
 import numpy as np
 import pytest
 import scipy.constants as constants
-from numpy.testing import assert_allclose
 from scipy.interpolate import interp2d
 
 import hyperspy.api as hs
 from hyperspy.datasets.example_signals import reference_hologram
+from hyperspy.decorators import lazifyTestClass
 
 
-@pytest.mark.parametrize('lazy', [True, False])
+@pytest.mark.parametrize("lazy", [True, False])
 def test_set_microscope_parameters(lazy):
-    holo_image = hs.signals.HologramImage(np.ones((3, 3)))
+    h = hs.signals.HologramImage(np.ones((3, 3)))
     if lazy:
-        holo_image = holo_image.as_lazy()
-    holo_image.set_microscope_parameters(
-        beam_energy=300., biprism_voltage=80.5, tilt_stage=2.2)
-    assert (holo_image.metadata.Acquisition_instrument.TEM.beam_energy == 300.)
-    assert (
-        holo_image.metadata.Acquisition_instrument.TEM.Biprism.voltage == 80.5)
-    assert (holo_image.metadata.Acquisition_instrument.TEM.Stage.tilt_alpha == 2.2)
+        h = h.as_lazy()
+
+    h.set_microscope_parameters(beam_energy=300.0, biprism_voltage=80.5, tilt_stage=2.2)
+
+    metadata = h.metadata.Acquisition_instrument.TEM
+    np.testing.assert_allclose(metadata.beam_energy, 300.0)
+    np.testing.assert_allclose(metadata.Biprism.voltage, 80.5)
+    np.testing.assert_allclose(metadata.Stage.tilt_alpha, 2.2)
+
+
+@lazifyTestClass
+class TestErrors:
+    def setup_method(self, method):
+        self.h = hs.signals.HologramImage(np.ones((5, 4)))
+
+    def test_absent_units(self):
+        with pytest.raises(ValueError, match="Signal axes units should be defined"):
+            self.h.statistics(sb_position=(1, 1))
+
+    def test_absent_beam_energy(self):
+        self.h.axes_manager.signal_axes[0].units = "nm"
+        self.h.axes_manager.signal_axes[1].units = "nm"
+
+        with pytest.raises(AttributeError, match="Please define the beam energy"):
+            self.h.statistics(sb_position=(1, 1))
+
+    def test_wrong_algorithm(self):
+        self.h.axes_manager.signal_axes[0].units = "nm"
+        self.h.axes_manager.signal_axes[1].units = "nm"
+        self.h.set_microscope_parameters(beam_energy=300.0)
+
+        with pytest.raises(ValueError, match="set to fourier or statistical"):
+            self.h.statistics(
+                sb_position=(1, 1), fringe_contrast_algorithm="pure_guess"
+            )
 
 
 def calc_holo(x, y, phase_ref, FRINGE_SPACING, FRINGE_DIRECTION):
-    return 2 * (1 + np.cos(phase_ref + FRINGE_SPACING / (2 * np.pi) * (
-        x * np.cos(FRINGE_DIRECTION) + y * np.sin(FRINGE_DIRECTION))))
+    mult = x * np.cos(FRINGE_DIRECTION) + y * np.sin(FRINGE_DIRECTION)
+    return 2 * (1 + np.cos(phase_ref + FRINGE_SPACING / (2 * np.pi) * mult))
 
 
 def calc_phaseref(x, y, z, img_sizex, img_sizey):
@@ -65,10 +93,8 @@ Y_START = int(IMG_SIZE3Y / 9)
 Y_STOP = IMG_SIZE3Y - 1 - int(IMG_SIZE3Y / 9)
 
 
-@pytest.mark.parametrize('parallel,lazy', [(True, False), (False, False),
-                                           (None, True)])
-def test_reconstruct_phase_single(parallel, lazy):
-
+@pytest.mark.parametrize("lazy", [True, False])
+def test_reconstruct_phase_single(lazy):
     x, y = np.meshgrid(LS, LS)
     phase_ref = calc_phaseref(x, 0, 0, img_size / 2.2, 1)
     holo = calc_holo(x, y, phase_ref, FRINGE_SPACING, FRINGE_DIRECTION)
@@ -78,45 +104,42 @@ def test_reconstruct_phase_single(parallel, lazy):
     if lazy:
         ref_image = ref_image.as_lazy()
         holo_image = holo_image.as_lazy()
-    wave_image = holo_image.reconstruct_phase(
-        ref, store_parameters=True, parallel=parallel)
-    sb_pos_cc = wave_image.metadata.Signal.Holography.Reconstruction_parameters.sb_position * (-1) + \
-        [img_size, img_size]
+    wave_image = holo_image.reconstruct_phase(ref, store_parameters=True)
 
-    sb_size_cc = wave_image.metadata.Signal.Holography.Reconstruction_parameters.sb_size
-    sb_smoothness_cc = wave_image.metadata.Signal.Holography.Reconstruction_parameters.sb_smoothness
-    sb_units_cc = wave_image.metadata.Signal.Holography.Reconstruction_parameters.sb_units
+    metadata = wave_image.metadata.Signal.Holography.Reconstruction_parameters
+    sb_pos_cc = metadata.sb_position * (-1) + [img_size, img_size]
+    sb_size_cc = metadata.sb_size
+    sb_smoothness_cc = metadata.sb_smoothness
+    sb_units_cc = metadata.sb_units
+
     wave_image_cc = holo_image.reconstruct_phase(
         ref_image,
         sb_position=sb_pos_cc,
         sb_size=sb_size_cc,
         sb_smoothness=sb_smoothness_cc,
         sb_unit=sb_units_cc,
-        parallel=parallel)
+    )
     X_START = int(wave_image.axes_manager.signal_shape[0] / 10)
     X_STOP = int(wave_image.axes_manager.signal_shape[0] * 9 / 10)
     wave_crop = wave_image.data[X_START:X_STOP, X_START:X_STOP]
     wave_cc_crop = wave_image_cc.data[X_START:X_STOP, X_START:X_STOP]
 
     # asserts that waves from different
-    assert_allclose(wave_crop, np.conj(wave_cc_crop), rtol=1e-3)
+    np.testing.assert_allclose(wave_crop, np.conj(wave_cc_crop), rtol=1e-3)
     # sidebands are complex conjugate; this also tests possibility of
     # reconstruction with given sideband parameters
 
     # Cropping the reconstructed and original phase images and comparing:
     X_START = int(img_size / 10)
     X_STOP = img_size - 1 - int(img_size / 10)
-    phase_new_crop = wave_image.unwrapped_phase(
-        parallel=parallel).data[X_START:X_STOP, X_START:X_STOP]
+    phase_new_crop = wave_image.unwrapped_phase().data[X_START:X_STOP, X_START:X_STOP]
     phase_ref_crop = phase_ref[X_START:X_STOP, X_START:X_STOP]
-    assert_allclose(phase_new_crop, phase_ref_crop, atol=0.02)
+    np.testing.assert_allclose(phase_new_crop, phase_ref_crop, atol=0.02)
 
 
-@pytest.mark.parametrize('parallel,lazy', [(True, False), (False, False),
-                                           (None, True)])
-def test_reconstruct_phase_nonstandard(parallel, lazy):
-    # 2. Testing reconstruction with non-standard output size for stacked
-    # images:
+@pytest.mark.parametrize("lazy", [True, False])
+def test_reconstruct_phase_nonstandard(lazy):
+    """Testing reconstruction with non-standard output size for stacked images"""
     gc.collect()
     x2, z2, y2 = np.meshgrid(LS, np.array([0, 1]), LS)
     phase_ref2 = calc_phaseref(x2, y2, z2, img_size / 2.2, img_size / 2.2)
@@ -130,21 +153,18 @@ def test_reconstruct_phase_nonstandard(parallel, lazy):
         ref_image2 = ref_image2.as_lazy()
         holo_image2 = holo_image2.as_lazy()
 
-    sb_position2 = ref_image2.estimate_sideband_position(
-        sb='upper', parallel=parallel)
-    sb_position2_lower = ref_image2.estimate_sideband_position(
-        sb='lower', parallel=parallel)
-    sb_position2_left = ref_image2.estimate_sideband_position(
-        sb='left', parallel=parallel)
-    sb_position2_right = ref_image2.estimate_sideband_position(
-        sb='right', parallel=parallel)
-    assert sb_position2 == sb_position2_left
-    assert sb_position2_lower == sb_position2_right
+    sb_position2 = ref_image2.estimate_sideband_position(sb="upper")
+    sb_position2_lower = ref_image2.estimate_sideband_position(sb="lower",)
+    sb_position2_left = ref_image2.estimate_sideband_position(sb="left",)
+    sb_position2_right = ref_image2.estimate_sideband_position(sb="right",)
+    np.testing.assert_allclose(sb_position2, sb_position2_left)
+    np.testing.assert_allclose(sb_position2_lower, sb_position2_right)
 
-    sb_size2 = ref_image2.estimate_sideband_size(
-        sb_position2, parallel=parallel)
-    output_shape = (np.int(sb_size2.inav[0].data * 2),
-                    np.int(sb_size2.inav[0].data * 2))
+    sb_size2 = ref_image2.estimate_sideband_size(sb_position2)
+    output_shape = (
+        np.int(sb_size2.inav[0].data * 2),
+        np.int(sb_size2.inav[0].data * 2),
+    )
 
     wave_image2 = holo_image2.reconstruct_phase(
         reference=ref_image2,
@@ -152,7 +172,7 @@ def test_reconstruct_phase_nonstandard(parallel, lazy):
         sb_size=sb_size2,
         sb_smoothness=sb_size2 * 0.05,
         output_shape=output_shape,
-        parallel=parallel)
+    )
     # a. Reconstruction with parameters provided as ndarrays should be
     # identical to above:
     wave_image2a = holo_image2.reconstruct_phase(
@@ -161,30 +181,28 @@ def test_reconstruct_phase_nonstandard(parallel, lazy):
         sb_size=sb_size2.data,
         sb_smoothness=sb_size2.data * 0.05,
         output_shape=output_shape,
-        parallel=parallel)
-    assert wave_image2 == wave_image2a
+    )
+    np.testing.assert_allclose(wave_image2, wave_image2a)
     del wave_image2a
     gc.collect()
     # interpolate reconstructed phase to compare with the input (reference
     # phase):
     interp_x = np.arange(output_shape[0])
     phase_interp0 = interp2d(
-        interp_x,
-        interp_x,
-        wave_image2.inav[0].unwrapped_phase(parallel=parallel).data,
-        kind='cubic')
+        interp_x, interp_x, wave_image2.inav[0].unwrapped_phase().data, kind="cubic",
+    )
     phase_new0 = phase_interp0(
         np.linspace(0, output_shape[0], img_size),
-        np.linspace(0, output_shape[0], img_size))
+        np.linspace(0, output_shape[0], img_size),
+    )
 
     phase_interp1 = interp2d(
-        interp_x,
-        interp_x,
-        wave_image2.inav[1].unwrapped_phase(parallel=parallel).data,
-        kind='cubic')
+        interp_x, interp_x, wave_image2.inav[1].unwrapped_phase().data, kind="cubic",
+    )
     phase_new1 = phase_interp1(
         np.linspace(0, output_shape[0], img_size),
-        np.linspace(0, output_shape[0], img_size))
+        np.linspace(0, output_shape[0], img_size),
+    )
 
     X_START = int(img_size / 10)
     X_STOP = img_size - 1 - int(img_size / 10)
@@ -192,23 +210,23 @@ def test_reconstruct_phase_nonstandard(parallel, lazy):
     phase_new_crop1 = phase_new1[X_START:X_STOP, X_START:X_STOP]
     phase_ref_crop0 = phase_ref2[0, X_START:X_STOP, X_START:X_STOP]
     phase_ref_crop1 = phase_ref2[1, X_START:X_STOP, X_START:X_STOP]
-    assert_allclose(phase_new_crop0, phase_ref_crop0, atol=0.05)
-    assert_allclose(phase_new_crop1, phase_ref_crop1, atol=0.04)
+    np.testing.assert_allclose(phase_new_crop0, phase_ref_crop0, atol=0.05)
+    np.testing.assert_allclose(phase_new_crop1, phase_ref_crop1, atol=0.04)
 
 
-@pytest.mark.parametrize('parallel,lazy', [(True, False), (False, False),
-                                           (None, True)])
-def test_reconstruct_phase_multi(parallel, lazy):
-
+@pytest.mark.parametrize("lazy", [True, False])
+def test_reconstruct_phase_multi(lazy):
     x3, z3, y3 = np.meshgrid(
         np.linspace(-IMG_SIZE3X / 2, IMG_SIZE3X / 2, IMG_SIZE3X),
-        np.arange(6), np.linspace(-IMG_SIZE3Y / 2, IMG_SIZE3Y / 2, IMG_SIZE3Y))
+        np.arange(6),
+        np.linspace(-IMG_SIZE3Y / 2, IMG_SIZE3Y / 2, IMG_SIZE3Y),
+    )
     phase_ref3 = calc_phaseref(x3, y3, z3 / 6, IMG_SIZE3X * 2, IMG_SIZE3Y * 2)
     newshape = (2, 3, IMG_SIZE3X, IMG_SIZE3Y)
-    holo3 = calc_holo(x3, y3, phase_ref3, FRINGE_SPACING3,
-                      FRINGE_DIRECTION3).reshape(newshape)
-    ref3 = calc_holo(x3, y3, 0, FRINGE_SPACING3,
-                     FRINGE_DIRECTION3).reshape(newshape)
+    holo3 = calc_holo(x3, y3, phase_ref3, FRINGE_SPACING3, FRINGE_DIRECTION3).reshape(
+        newshape
+    )
+    ref3 = calc_holo(x3, y3, 0, FRINGE_SPACING3, FRINGE_DIRECTION3).reshape(newshape)
     del x3, z3, y3
     gc.collect()
     holo_image3 = hs.signals.HologramImage(holo3)
@@ -218,32 +236,35 @@ def test_reconstruct_phase_multi(parallel, lazy):
         ref_image3 = ref_image3.as_lazy()
         holo_image3 = holo_image3.as_lazy()
 
-    wave_image3 = holo_image3.reconstruct_phase(
-        ref_image3.inav[0, 0], sb='upper', parallel=parallel)
+    wave_image3 = holo_image3.reconstruct_phase(ref_image3.inav[0, 0], sb="upper",)
 
     # Cropping the reconstructed and original phase images and comparing:
-    phase3_new_crop = wave_image3.unwrapped_phase(parallel=parallel)
+    phase3_new_crop = wave_image3.unwrapped_phase()
     phase3_new_crop.crop(2, Y_START, Y_STOP)
     phase3_new_crop.crop(3, X_START, X_STOP)
-    phase3_ref_crop = phase_ref3.reshape(newshape)[:, :, X_START:X_STOP,
-                                                   Y_START:Y_STOP]
-    assert_allclose(phase3_new_crop.data, phase3_ref_crop, atol=0.7)
+    phase3_ref_crop = phase_ref3.reshape(newshape)[:, :, X_START:X_STOP, Y_START:Y_STOP]
+    np.testing.assert_allclose(phase3_new_crop.data, phase3_ref_crop, atol=0.7)
 
     # 3a. Testing reconstruction with input parameters in 'nm' and with multiple parameter input,
     # but reference ndim=0:
 
-    sb_position3 = ref_image3.estimate_sideband_position(
-        sb='upper', parallel=parallel)
-    f_sampling = np.divide(1, [
-        a * b
-        for a, b in zip(ref_image3.axes_manager.signal_shape,
-                        (ref_image3.axes_manager.signal_axes[0].scale,
-                         ref_image3.axes_manager.signal_axes[1].scale))
-    ])
-    sb_size3 = ref_image3.estimate_sideband_size(
-        sb_position3, parallel=parallel) * np.mean(f_sampling)
+    sb_position3 = ref_image3.estimate_sideband_position(sb="upper")
+    f_sampling = np.divide(
+        1,
+        [
+            a * b
+            for a, b in zip(
+                ref_image3.axes_manager.signal_shape,
+                (
+                    ref_image3.axes_manager.signal_axes[0].scale,
+                    ref_image3.axes_manager.signal_axes[1].scale,
+                ),
+            )
+        ],
+    )
+    sb_size3 = ref_image3.estimate_sideband_size(sb_position3,) * np.mean(f_sampling)
     sb_smoothness3 = sb_size3 * 0.05
-    sb_units3 = 'nm'
+    sb_units3 = "nm"
 
     wave_image3a = holo_image3.reconstruct_phase(
         ref_image3.inav[0, 0],
@@ -251,185 +272,164 @@ def test_reconstruct_phase_multi(parallel, lazy):
         sb_size=sb_size3,
         sb_smoothness=sb_smoothness3,
         sb_unit=sb_units3,
-        parallel=parallel)
-    phase3a_new_crop = wave_image3a.unwrapped_phase(parallel=parallel)
+    )
+    phase3a_new_crop = wave_image3a.unwrapped_phase()
     phase3a_new_crop.crop(2, Y_START, Y_STOP)
     phase3a_new_crop.crop(3, X_START, X_STOP)
-    assert_allclose(phase3a_new_crop.data, phase3_ref_crop, atol=0.7)
+    np.testing.assert_allclose(phase3a_new_crop.data, phase3_ref_crop, atol=0.7)
+
     # a. Mismatch of navigation dimensions of object and reference
     # holograms, except if reference hologram ndim=0
     with pytest.raises(ValueError):
         holo_image3.reconstruct_phase(ref_image3.inav[0, :])
     reference4a = ref_image3.inav[0, :]
-    reference4a.set_signal_type('signal2d')
+    reference4a.set_signal_type("signal2d")
     with pytest.raises(ValueError):
         holo_image3.reconstruct_phase(reference=reference4a)
     #   b. Mismatch of signal shapes of object and reference holograms
     with pytest.raises(ValueError):
         holo_image3.reconstruct_phase(
-            ref_image3.inav[:, :].isig[Y_START:Y_STOP, X_START:X_STOP])
+            ref_image3.inav[:, :].isig[Y_START:Y_STOP, X_START:X_STOP]
+        )
 
     #   c. Mismatch of signal shape of sb_position
     sb_position_mismatched = hs.signals.Signal2D(np.arange(9).reshape((3, 3)))
     with pytest.raises(ValueError):
-        holo_image3.reconstruct_phase(
-            sb_position=sb_position_mismatched)
+        holo_image3.reconstruct_phase(sb_position=sb_position_mismatched)
     #   d. Mismatch of navigation dimensions of reconstruction parameters
     sb_position_mismatched = hs.signals.Signal1D(np.arange(16).reshape((8, 2)))
     sb_size_mismatched = hs.signals.BaseSignal(np.arange(9)).T
     sb_smoothness_mismatched = hs.signals.BaseSignal(np.arange(9)).T
     with pytest.raises(ValueError):
-        holo_image3.reconstruct_phase(
-            sb_position=sb_position_mismatched)
+        holo_image3.reconstruct_phase(sb_position=sb_position_mismatched)
     with pytest.raises(ValueError):
-        holo_image3.reconstruct_phase(
-            sb_size=sb_size_mismatched)
+        holo_image3.reconstruct_phase(sb_size=sb_size_mismatched)
     with pytest.raises(ValueError):
-        holo_image3.reconstruct_phase(
-            sb_smoothness=sb_smoothness_mismatched)
+        holo_image3.reconstruct_phase(sb_smoothness=sb_smoothness_mismatched)
 
     #   e. Beam energy is not assigned, while 'mrad' units selected
     with pytest.raises(AttributeError):
-        holo_image3.reconstruct_phase(
-            sb_size=40, sb_unit='mrad')
+        holo_image3.reconstruct_phase(sb_size=40, sb_unit="mrad")
 
 
-def _generate_parameters():
-    parameters = []
-    for parallel in [False, True]:
-        for lazy in [False, True]:
-            for single_values in [False, True]:
-                for fcalgo in ['fourier', 'statistical']:
-                    parameters.append([parallel,
-                                       lazy,
-                                       single_values,
-                                       fcalgo])
-    return parameters
+class TestStatistics:
+    def setup_method(self, method):
+        # Parameters measured using Gatan HoloWorks:
+        self.REF_FRINGE_SPACING = 3.48604
+        self.REF_FRINGE_SAMPLING = 3.7902
 
+        # Measured using the definition of fringe contrast from the centre of image
+        self.REF_FRINGE_CONTRAST = 0.0736
 
-@pytest.mark.parametrize('parallel, lazy, single_values, fringe_contrast_algorithm',
-                         _generate_parameters())
-def test_statistics(parallel, lazy, single_values, fringe_contrast_algorithm):
-    # Parameters measured using Gatan HoloWorks:
-    REF_FRINGE_SPACING = 3.48604
-    REF_FRINGE_SAMPLING = 3.7902
+        # Prepare test data and derived statistical parameters
+        self.ref_carrier_freq = 1.0 / self.REF_FRINGE_SAMPLING
+        self.ref_carrier_freq_nm = 1.0 / self.REF_FRINGE_SPACING
 
-    # Measured using the definition of fringe contrast from the centre of image
-    REF_FRINGE_CONTRAST = 0.0736
+        # 3x2 stack
+        self.ref_holo = hs.stack([reference_hologram()] * 2)
+        self.ref_holo = hs.stack([self.ref_holo] * 3)
 
-    RTOL = 1e-5
+        ht = self.ref_holo.metadata.Acquisition_instrument.TEM.beam_energy
+        momentum = (
+            2
+            * constants.m_e
+            * constants.elementary_charge
+            * ht
+            * 1000
+            * (
+                1
+                + constants.elementary_charge
+                * ht
+                * 1000
+                / (2 * constants.m_e * constants.c ** 2)
+            )
+        )
+        wavelength = constants.h / np.sqrt(momentum) * 1e9  # in nm
+        self.ref_carrier_freq_mrad = self.ref_carrier_freq_nm * 1000 * wavelength
 
-    # 0. Prepare test data and derived statistical parameters
+    @pytest.mark.parametrize("lazy", [True, False])
+    @pytest.mark.parametrize("fringe_contrast_algorithm", ["fourier", "statistical"])
+    def test_single_values(self, lazy, fringe_contrast_algorithm):
+        if lazy:
+            self.ref_holo.as_lazy()
 
-    ref_carrier_freq = 1. / REF_FRINGE_SAMPLING
-    ref_carrier_freq_nm = 1. / REF_FRINGE_SPACING
-
-    ref_holo = hs.stack([reference_hologram(), reference_hologram()])
-    ref_holo = hs.stack([ref_holo, ref_holo, ref_holo])
-
-    ht = ref_holo.metadata.Acquisition_instrument.TEM.beam_energy
-    momentum = 2 * constants.m_e * constants.elementary_charge * ht * \
-        1000 * (1 + constants.elementary_charge * ht *
-                1000 / (2 * constants.m_e * constants.c ** 2))
-    wavelength = constants.h / np.sqrt(momentum) * 1e9  # in nm
-    ref_carrier_freq_mrad = ref_carrier_freq_nm * 1000 * wavelength
-
-    if lazy:
-        ref_holo.as_lazy()
-
-    # 1. Test core functionality
-
-    stats = ref_holo.statistics(high_cf=True,
-                                parallel=parallel,
-                                single_values=single_values,
-                                fringe_contrast_algorithm=fringe_contrast_algorithm)
-    if single_values:
-        # Fringe contrast in experimental conditions can be only an estimate
-        # therefore tolerance is 10%:
-        assert_allclose(
-            stats['Fringe contrast'],
-            REF_FRINGE_CONTRAST,
-            rtol=0.1)
-
-        assert_allclose(
-            stats['Fringe sampling (px)'],
-            REF_FRINGE_SAMPLING,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Fringe spacing (nm)'],
-            REF_FRINGE_SPACING,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Carrier frequency (1 / nm)'],
-            ref_carrier_freq_nm,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Carrier frequency (1/px)'],
-            ref_carrier_freq,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Carrier frequency (mrad)'],
-            ref_carrier_freq_mrad,
-            rtol=RTOL)
-    else:
-        ref_fringe_contrast_stack = np.repeat(
-            REF_FRINGE_CONTRAST, 6).reshape((3, 2))
-        ref_fringe_sampling_stack = np.repeat(
-            REF_FRINGE_SAMPLING, 6).reshape((3, 2))
-        ref_fringe_spacing_stack = np.repeat(
-            REF_FRINGE_SPACING, 6).reshape((3, 2))
-        ref_carrier_freq_nm_stack = np.repeat(
-            ref_carrier_freq_nm, 6).reshape((3, 2))
-        ref_carrier_freq_stack = np.repeat(ref_carrier_freq, 6).reshape((3, 2))
-        ref_carrier_freq_mrad_stack = np.repeat(
-            ref_carrier_freq_mrad, 6).reshape((3, 2))
+        stats = self.ref_holo.statistics(
+            high_cf=True,
+            single_values=True,
+            fringe_contrast_algorithm=fringe_contrast_algorithm,
+        )
 
         # Fringe contrast in experimental conditions can be only an estimate
         # therefore tolerance is 10%:
-        assert_allclose(
-            stats['Fringe contrast'].data,
-            ref_fringe_contrast_stack,
-            rtol=0.1)
+        np.testing.assert_allclose(
+            stats["Fringe contrast"], self.REF_FRINGE_CONTRAST, rtol=0.1
+        )
 
-        assert_allclose(
-            stats['Fringe sampling (px)'].data,
-            ref_fringe_sampling_stack,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Fringe spacing (nm)'].data,
-            ref_fringe_spacing_stack,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Carrier frequency (1 / nm)'].data,
+        np.testing.assert_allclose(
+            stats["Fringe sampling (px)"], self.REF_FRINGE_SAMPLING, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Fringe spacing (nm)"], self.REF_FRINGE_SPACING, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Carrier frequency (1 / nm)"], self.ref_carrier_freq_nm, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Carrier frequency (1/px)"], self.ref_carrier_freq, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Carrier frequency (mrad)"], self.ref_carrier_freq_mrad, rtol=1e-5
+        )
+
+    @pytest.mark.parametrize("lazy", [True, False])
+    @pytest.mark.parametrize("fringe_contrast_algorithm", ["fourier", "statistical"])
+    def test_no_single_values(self, lazy, fringe_contrast_algorithm):
+        if lazy:
+            self.ref_holo.as_lazy()
+
+        stats = self.ref_holo.statistics(
+            high_cf=True,
+            single_values=False,
+            fringe_contrast_algorithm=fringe_contrast_algorithm,
+        )
+
+        ref_fringe_contrast_stack = np.repeat(self.REF_FRINGE_CONTRAST, 6).reshape(
+            (3, 2)
+        )
+        ref_fringe_sampling_stack = np.repeat(self.REF_FRINGE_SAMPLING, 6).reshape(
+            (3, 2)
+        )
+        ref_fringe_spacing_stack = np.repeat(self.REF_FRINGE_SPACING, 6).reshape((3, 2))
+        ref_carrier_freq_nm_stack = np.repeat(self.ref_carrier_freq_nm, 6).reshape(
+            (3, 2)
+        )
+        ref_carrier_freq_stack = np.repeat(self.ref_carrier_freq, 6).reshape((3, 2))
+        ref_carrier_freq_mrad_stack = np.repeat(self.ref_carrier_freq_mrad, 6).reshape(
+            (3, 2)
+        )
+
+        # Fringe contrast in experimental conditions can be only an estimate
+        # therefore tolerance is 10%:
+        np.testing.assert_allclose(
+            stats["Fringe contrast"].data, ref_fringe_contrast_stack, rtol=0.1
+        )
+
+        np.testing.assert_allclose(
+            stats["Fringe sampling (px)"].data, ref_fringe_sampling_stack, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Fringe spacing (nm)"].data, ref_fringe_spacing_stack, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Carrier frequency (1 / nm)"].data,
             ref_carrier_freq_nm_stack,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Carrier frequency (1/px)'].data,
-            ref_carrier_freq_stack,
-            rtol=RTOL)
-        assert_allclose(
-            stats['Carrier frequency (mrad)'].data,
+            rtol=1e-5,
+        )
+        np.testing.assert_allclose(
+            stats["Carrier frequency (1/px)"].data, ref_carrier_freq_stack, rtol=1e-5
+        )
+        np.testing.assert_allclose(
+            stats["Carrier frequency (mrad)"].data,
             ref_carrier_freq_mrad_stack,
-            rtol=RTOL)
-
-    # 2. Test raises:
-    holo_raise = hs.signals.HologramImage(np.random.random(20).reshape((5, 4)))
-
-    # 2a. Test raise for absent units:
-    with pytest.raises(ValueError):
-        holo_raise.statistics(sb_position=(1, 1))
-    holo_raise.axes_manager.signal_axes[0].units = 'nm'
-    holo_raise.axes_manager.signal_axes[1].units = 'nm'
-
-    # 2b. Test raise for absent beam_energy:
-    with pytest.raises(AttributeError):
-        holo_raise.statistics(sb_position=(1, 1))
-    holo_raise.set_microscope_parameters(beam_energy=300.)
-
-    # 2c. Test raise for wrong value of `fringe_contrast_algorithm`
-    with pytest.raises(ValueError):
-        holo_raise.statistics(
-            sb_position=(
-                1,
-                1),
-            fringe_contrast_algorithm='pure_guess')
+            rtol=1e-5,
+        )
